refactor: extract summary_utils.py from scraper.py
Move 10 module-level utility functions (~200 lines) to summary_utils.py.
These are pure functions used only for run summary building and result
flattening. Re-export merge_context and flatten_results from scraper.py
for backward compatibility. Update test imports accordingly.
